Shopping page updated
"Test" user has been used to test the link. Once logged in Shopping page opens and you can add items to the cart and view cart. At some point the search feature got broken and need to be fixed. Working on linking user Cart to Postgres database. Item quantity can be increased by clicking on the "add to cart" button multiple times. The images are local on my computer for the items I'll add a folder in the repo with those images in case anyone wants to test it themselves.  Will start on check out after search is fixed and cart is saving to database.